requirements.txt
fastapi==0.115.12
uvicorn==0.34.0
sse_starlette==2.2.1
slowapi==0.1.9
httpx==0.28.1
langfuse>=2,<3
litellm>=1.70.0